server: better shutdown, and faster exit on *nix
TL;DR: on *nix, shutdown now start immediately instead of after 3 seconds
and otherwise behaves identically as before (which means it can hang if
shutdown fails - as before). On Windows, behaves identically except the
timeout changed from 3s to 5s. On all tested OSs it seems to always shutdown
successfully and quickly (within ~200ms).

The shutdown code at mwServerShutdown assumed it can tell the http loop
to shutdown and then wait and poll it till it's done. That was correct
for Windows but not for linux.

The reason is that the http loop runs on the main thread, and on Windows
ctrl-c goes via the console control handler which runs on a different thread,
so it can wait till the loop exits.

On *nix, however, the call to shutdown is triggered via the signal handler,
which is the same thread as main, so as long as the handler "waits" for the
loop, the loop cannot progress.

The solution here is to have mwServerShutdown act differently depending on
which thread its running on, indicated by a new callback argument - if the
callback exists it means the caller is on the same thread as the server, and
it will signal the loop to exit but return right away, and once control
resumes to the server after the handler exits, the loop will try to exit and
then call the callback - successful shutdown is not guaranteed.

Currently there's no system to do anything if the callback is not invoked,
but it's better than before since it was waiting 3s before the actual
shutdown started, and now the shutdown starts right away.

On windows there's no change as it was working, except the timeout was raised
from 3000ms to 5000ms.