remove buggy addr info display
When a specific interface/address is _not_ given with -i, it binds to all
interfaces (addr 0), however, it only displayed the address of the first
interface, and only if the hostname can be resolved (which is not always
the case even if it should be) or else it crashed.

The display code has no effect on the actual binding or operation of the
server.

So remove that buggy code and instead display "0.0.0.0" on such case.

In the future we might change it to actually display all addresses.