URL decoding: fail explicitlly on %00 or invalid %HH escape
%00 is a valid escape but not a valid UTF8 middle-byte and we don't
support url arguments anyway, so it's not a valid part of a URL for us.

Additionally, if a request URI ended with '%' followed by one char,
then the code which resolves the local file name was reading past the
end of input string. This was fixed implicitly due to the error checks.