request URI: completely disallow '.' and '..' components
This is a rewrite of mwGetLocalFileName(...) .

Previously the code was shorter, but allowed some cases of meta-dirs
to be missed, e.g. http://my.domain/.%2e/ would have served one dir
above (outside) the web root dir.

Additionally, the request was silently truncated if it was too long.

Now we take the longer approach: first decode URL escapes ('+' and %HH),
then convert and coallesc all slash or backslash to our standard slash
form, then check that we don't have meta dirs . or .. in it, and finally
reject if it's too long.