fix directory listing when the server root path is /
When the root path of the server is "/" and the request is "/", the
logic which converts the request to local path ended up empty (after
trimming the trailing slash and marking it as if it had trailing /).

But "/" is a special case which shouldn't be trimmed.

This commit keeps "/" on such case, and modify few other places which
didn't expect the dir name to end with "/".